requirements.txt
influxdb
jsonschema
pyyaml
simplepush